refactor: Piece 인터페이스와 PieceType을 piece 패키지로 이동
- Piece, PieceType이 도메인 루트에 있던 것을 piece 패키지로 응집
- piece 패키지가 자신의 인터페이스와 타입을 직접 관리하도록 개선
