+++ src/main/java/janggi/domain/board/Board.java
@@ -1,6 +1,6 @@
 package janggi.domain.board;
 
-import janggi.domain.Piece;
+import janggi.domain.piece.Piece;
 import janggi.domain.piece.EmptyPiece;
 import janggi.domain.position.Position;
 import java.util.ArrayList;
