+++ src/main/java/janggi/domain/piece/PieceType.java
@@ -1,4 +1,4 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 public enum PieceType {
     CHARIOT("차", "CHA"),
