+++ src/test/java/janggi/domain/board/BoardFactoryTest.java
@@ -2,7 +2,7 @@ package janggi.domain.board;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import janggi.domain.Piece;
+import janggi.domain.piece.Piece;
 import janggi.domain.Team;
 import janggi.domain.piece.Cannon;
 import janggi.domain.piece.Chariot;
